requirements.txt
flask==2.0.1
werkzeug==2.0.3
ezdxf==1.0.0
reportlab==3.6.8
Pillow==10.2.0
python-dotenv==1.0.0
matplotlib==3.5.1 